split library into modules
